core.common: move assert_subpackage to my.core.internal
