+++ my/core/mime.py
@@ -2,7 +2,7 @@
 Utils for mime/filetype handling
 """
 
-from .common import assert_subpackage; assert_subpackage(__name__)
+from .internal import assert_subpackage; assert_subpackage(__name__)
 
 import functools
