+++ my/core/pytest.py
@@ -2,7 +2,7 @@
 Helpers to prevent depending on pytest in runtime
 """
 
-from .common import assert_subpackage; assert_subpackage(__name__)
+from .internal import assert_subpackage; assert_subpackage(__name__)
 
 import sys
 import typing
